- Remove unused and old files. Reorganized Python DB/REST API utilities
