--- example/dbapi
@@ -1,1 +1,0 @@
-../util/dbapi
