--- util/.flake8
@@ -1,3 +1,0 @@
-[flake8]
-max-line-length = 132
-
