--- util/README
@@ -1,59 +1,0 @@
-Countries missing from the country data
----------------------------------------
-ID 106 - Finland
-ID 108 - Falklands
-ID 109 - France
-ID 111 - UK
-ID 119 - Greenland
-ID 133 - Israel
-ID 169 - Vanatu
-ID 173 - Nederland
-ID 176 - New Zealand
-ID 184 - Puerto Rico
-ID 223 - USA
-ID 233 - Aruba Island
-ID 234 - Anguilla Island (Antilies)
-ID 235 - Finland ?
-ID 237 - Samoan Islands
-ID 242 - France
-ID 244 - Austria
-ID 246 - Rarotonga (Cook Islands)
-ID 250 - Curaçao Island (Antilies)
-ID 251 - Cayman Islands
-ID 253 - Cyprus
-ID 254 - Faroe Islands (Denmark)
-ID 256 - Guernsey Island
-ID 257 - Gibraltar
-ID 259 - Guam
-ID 260 - Hong Kong
-ID 261 - Heard Island and McDonald Islands (Austrailia)
-ID 262 - Isle of Man
-ID 263 - Cocos (Keeling) Islands (Austrailia)
-ID 264 - Chagos Archipelago
-ID 265 - Jersey Island
-ID 266 - Kazakhstan
-ID 272 - Macau
-ID 273 - Île des Terres Basses Island (Antitles)
-ID 278 - Rota Island (Mariana Islands)
-ID 279 - Montserrat Island
-ID 281 - Norfolk Island (Australia)
-ID 282 - Koloa Island (Tonga)
-ID 284 - Pitcairn Islands
-ID 287 - West Bank (Palestine/Israel)
-ID 288 - Tahiti
-ID 292 - South Georgia
-ID 293 - St Helena
-ID 296 - Saint Pierre Island
-ID 298 - Country of Sint Maarten
-ID 300 - Turks and Caicos Islands 
-ID 303 - Johnston Island
-ID 304 - Cuba
-ID 307 - British Virgin Islands
-ID 308 - Saint John
-ID 309 - Île Futuna
-ID 310 - Cyprus
-ID 60 - Possession Island and East Island
-ID 61 - Australia
-ID 74 - Pembroke Island, Bermuda
-ID 84 - China ?
-ID 97 - Zealand (Denmark)
