+++ util/python/dbapi/example/csv-report.py
@@ -1,0 +1,48 @@
+#!/usr/bin/python3
+#
+# Copyright (c) 2023 Humanitarian OpenStreetMap Team
+#
+# This file is part of Underpass.
+#
+#     Underpass is free software: you can redistribute it and/or modify
+#     it under the terms of the GNU General Public License as published by
+#     the Free Software Foundation, either version 3 of the License, or
+#     (at your option) any later version.
+#
+#     Underpass is distributed in the hope that it will be useful,
+#     but WITHOUT ANY WARRANTY; without even the implied warranty of
+#     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#     GNU General Public License for more details.
+#
+#     You should have received a copy of the GNU General Public License
+#     along with Underpass.  If not, see <https://www.gnu.org/licenses/>.
+
+#     This is an example on how you can use the Underpass API
+#     to get data from the Underpass database
+#
+#     Install psycopg2 (pip install psycopg2) and run (python report.py)
+
+import sys,os
+sys.path.append(os.path.realpath('..'))
+
+from api import report
+
+reporter = report.Report()
+reporter.responseType = "csv"
+
+results = reporter.getDataQualityTag(
+    fromDate = "2023-03-01T00:00:00", 
+    toDate = "2023-03-01T23:59:59", 
+    hashtags = ["hotosm"],
+    area = "MULTIPOLYGON(((35.94239853552784 34.604923785746294,36.387327544252344 34.60572420609614,36.619489146712056 34.27201098162725,35.98750002858665 33.796472610236506,35.80019372486123 33.371269706837424,35.84379302937904 32.945245369785624,35.7822480812834 32.677308476658354,36.86461568162349 32.3013811095846,40.772406757157626 34.31327058449696,41.251362560106344 34.80363125920435,41.26519612281126 35.472947952543976,41.39023183948342 35.73126442006469,41.390598531061926 36.45206157821798,42.56115535177841 37.04978378309069,42.33115259542802 37.290342798776564,43.64022201603803 37.27592913492714,44.21186544642592 37.171051186202305,44.22726396563035 36.88504777495011,44.79595017217301 37.1338359471719,44.67270121531499 37.67818055969157,44.22580546785684 37.90139699481624,44.56513289411458 38.32915934737434,44.11870988528574 39.30468107118455,44.77128751703804 39.778182229320294,44.251905891863345 40.075458922584176,43.72400891687824 40.135132989679306,43.789093422415135 40.933345200697204,42.93051401810241 41.477544053322305,42.43971449070102 41.47465105160185,41.543056142787634 41.49982852812053,40.345523155310865 40.965130053155235,39.36516907401443 41.12773635772109,38.206216060282635 40.90864908433082,37.25575551744291 41.1494297110911,36.76660091198434 41.31336930155078,36.368395877105456 41.34992938569337,36.03865175548873 41.60302214675281,35.82540667767503 41.71148619328122,35.37687367859127 41.691690090750456,35.108504832585226 42.01819317235436,34.632226289772916 42.01489868742976,33.21995228166264 42.03552198939525,31.431973181497312 41.30648453380394,30.999033182953895 41.07682438034638,29.18763785805726 41.16405992385131,28.056996710112628 41.584316257808126,28.00787734802205 41.89647214891835,27.539507226074534 41.9059550085305,27.279221051301047 42.12012016868917,26.422943279605647 41.87474259521338,26.26871517199308 41.75874532609109,26.573228021002596 41.33749460107572,26.25414585117366 41.03442077977786,26.10429650958611 40.75989762187643,26.175863662305517 40.063015367675575,25.859797941336097 39.44152441767588,25.608725498376828 38.699587739879775,25.722853166187463 37.918714381749695,26.17743471708161 37.18837513594946,26.629030482960218 36.52809264167125,27.391260784075484 35.80874375489056,28.725320259563578 35.78879474946699,31.13914781136154 36.114080099972455,33.09666408666283 35.739303068700295,35.504835678237896 36.358959428941105,35.94239853552784 34.604923785746294)))"
+)
+
+print(results)
+
+# results = reporter.getDataQualityGeo(
+#     fromDate = "2023-03-01T00:00:00", 
+#     toDate = "2023-03-01T23:59:59", 
+#     hashtags = ["hotosm", "turkey"],
+#     page = 3
+# )
+
