--- util/requirements.txt
@@ -1,6 +1,0 @@
-codetiming==1.3.0
-epdb==0.15.1
-pandas==1.3.4
-progress==1.6
-psycopg2==2.9.1
-SQLAlchemy==1.4.27
